app/requirements.txt
Flask 
pylint 
pandas
wikipedia
textblob
google-cloud-language
google-cloud-bigquery
flasgger

pandas
sensible
click
pytest
pylint
ipython
flask
pylint-flask
Flask-API
requests
flasgger
pytest-flask
pytest-cov
matplotlib
seaborn
numpy